+++ src/vatensor/comparison.hpp
@@ -12,6 +12,10 @@ namespace va {
 	void less_equal(VStoreAllocator& allocator, VArrayTarget target, const VData& a, const VData& b);
 	void is_close(VStoreAllocator& allocator, VArrayTarget target, const VData& a, const VData& b, double rtol = 1e-05, double atol = 1e-08, bool equal_nan = false);
 
+	void is_nan(VStoreAllocator& allocator, VArrayTarget target, const VData& a);
+	void is_finite(VStoreAllocator& allocator, VArrayTarget target, const VData& a);
+	void is_inf(VStoreAllocator& allocator, VArrayTarget target, const VData& a);
+
 	bool array_equal(const VData& a, const VData& b);
 	bool all_close(const VData& a, const VData& b, double rtol = 1e-05, double atol = 1e-08, bool equal_nan = false);
 }